Add basic block and board structs.
The Block struct has a colour, and has a method to draw itself (by
drawing a square the correct colour).

The Board struct has a 2d grid, each element of which is an
Option<Block>. This is either the block at that space on the board, or
None if there is no block at that space. The board has a draw method
which iterates each square on the grid and draws the block at that
space, if there is one.