Add tetromino struct
Struct contains an X and Y position for the tetromino, and an array of
four blocks. The blocks themselves each have an x and y position, within
a 4x4 grid. Tetromino x + block x gives the absolute x position of a
block within the game grid.

There is an enumeration of all different shapes, and the constructor
takes this as its argument. The block positions and colours are
initialised based on this value.

There is a draw method that works by drawing each of the blocks.