Add tetris::Game struct
Ties together a board and an active tetromino. Responds to input events
and time by moving the tetromino. When a tetromino cannot move down it
gets added to the board and a new tetromino is randomly generated.

Blocks are not cleared when a row is filled. This will crash when the
screen fills rather than gracefully handling a game over.